+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - URL path changed
     - `/hash` -> [`/hash-text`](https://it-tools.tech/hash-text)
     - `/cypher` -> [`/cypher-uncyfer-text`](https://it-tools.tech/cypher-uncyfer-text)
+    - `/url-encoder` -> [`/url-encoder-decoder`](https://it-tools.tech/url-encoder-decoder)
 - [Crontab generator](https://it-tools.tech/crontab-generator) now handle 6 digits format (second support)  
 
 ### Added